LayoutTests/media/video-seek-past-end-paused-expected.txt

Test that seeking paused video past it's duration time sets currentTime to duration and leaves video paused.

RUN(video.load())
EVENT(load)
RUN(video.currentTime = 500)
EVENT(timeupdate)
EXPECTED (video.currentTime == '8.399999618530273') OK
END OF TEST
